fix: use u8 instead of u64

+++ src/_03/mod.rs
@@ -1,4 +1,4 @@
-type Battery = u64;
+type Battery = u8;
 type Bank = Vec<Battery>;
 
 fn parse(i: &'static str) -> Vec<Bank> {
